# docs/swagger.yaml
definitions:
  ent.User:
    properties:
      avatar:
        description: 头像路径
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      created_by:
        description: 创建者
        type: string
      deleted_time:
        description: 删除标志（0代表存在）
        type: integer
      display_name:
        description: 显示名称
        type: string
      email:
        description: 用户邮箱
        type: string
      gender:
        description: 用户性别（0保密 1男 2女）
        type: string
      id:
        description: |-
          ID of the ent.
          用户ID
        type: integer
      name:
        description: 用户账号
        type: string
      non_locked:
        description: 锁定标志（0锁定 1正常）
        type: integer
      phone:
        description: 手机号码
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      updated_by:
        description: 更新者
        type: string
    type: object
  types.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  types.LoginOutput:
    properties:
      token:
        type: string
    type: object
  types.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /account/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.LoginOutput'
      summary: 账号登录
      tags:
      - account
  /account/profile:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
      summary: 账号信息
      tags:
      - account
  /account/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.User'
      summary: 账号注册
      tags:
      - account
swagger: "2.0"